+++ src/main/java/net/geforcemods/securitycraft/network/server/SetBriefcasePasscodeAndOwner.java
@@ -4,14 +4,17 @@ import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.components.OwnerData;
 import net.geforcemods.securitycraft.components.PasscodeData;
+import net.geforcemods.securitycraft.network.client.OpenScreen;
 import net.geforcemods.securitycraft.util.PasscodeUtils;
 import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.neoforged.neoforge.network.PacketDistributor;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 
 public record SetBriefcasePasscodeAndOwner(String passcode) implements CustomPacketPayload {
@@ -35,8 +38,11 @@ public record SetBriefcasePasscodeAndOwner(String passcode) implements CustomPac
 			if (!stack.has(SCContent.OWNER_DATA))
 				stack.set(SCContent.OWNER_DATA, OwnerData.fromPlayer(player, true));
 
-			if (!passcode.isEmpty() && !stack.has(SCContent.PASSCODE_DATA))
-				PasscodeData.hashAndSetPasscode(stack, passcode, newPasscodeData -> {});
+			if (!passcode.isEmpty() && !stack.has(SCContent.PASSCODE_DATA)) {
+				PasscodeData.hashAndSetPasscode(stack, passcode, newPasscodeData -> {
+					PacketDistributor.sendToPlayer((ServerPlayer) player, new OpenScreen(OpenScreen.DataType.CHECK_PASSCODE_FOR_BRIEFCASE));
+				});
+			}
 		}
 	}
 }